kloud: move route53 package to a independent package
Useful for creating tests around it
